+++ Assets/Scripts/Player/PlayerStatistic.cs
@@ -5,8 +5,8 @@ using UnityEngine;
 public class PlayerStatistic : MonoBehaviour
 {
     public static PlayerStatistic Instance;
-    private int life;
-    public int Life { get { return life; } set { life = Mathf.Clamp(value, 0, 100); } }
+    private float life;
+    public float Life { get { return life; } set { life = Mathf.Clamp(value, 0, 100); } }
     public int Attack { get; set; }
     public int Armor_Helmet { get; set; }
     public int Armor_Chestplate { get; set; }
